requirements.txt
fastapi==0.120.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-dotenv==1.0.0

# Development dependencies
ruff>=0.1.0

